+++ src/axios/inquiry/Inquiry.js
@@ -34,3 +34,73 @@ export function cancelOrder(dto, receiverPhoneNumber) {
       }
     });
 }
+
+export function fetchRealTimeDeliveryInfo(orderId, phoneNumber, orderDetails, popup) {
+  const orderDelivery = orderDetails.orderDelivery;
+  const courierCode = orderDelivery.code;
+  const trackingNumber = orderDelivery.trackingNumber;
+
+  const getAccessToken = () => {
+    return sessionStorage.getItem("ACCESS_TOKEN");
+  }
+
+  const fetchUserDeliveryInfo = (accessToken, orderId) => {
+      axios.get(`/product/orders/${orderId}/delivery?courierCode=${courierCode}&trackingNumber=${trackingNumber}`, {
+        headers: {
+          Authorization: `${accessToken}`,
+        },
+      })
+      .then(response => {
+        if (response.status === 200 || response.status === 302) {
+          if (!popup) {
+            alert("팝업 차단을 해제해주세요")
+          } else {
+            popup.location.href = response.request?.responseURL;
+          }
+        }
+         else {
+          const data = response.json();
+          data.then((res => {alert(res.errorName)}));
+        }
+      })
+      .catch(error => {
+        console.error(error);
+      });
+  }
+
+  const fetchGuestDeliveryInfo = (phoneNumber, orderNumber) => {
+    axios.get(`/guest/product/orders/${orderNumber}/delivery?courierCode=${courierCode}&trackingNumber=${trackingNumber}`, {
+      headers: {
+        Authorization: `${phoneNumber}`,
+      },
+    })
+    .then(response => {
+      if (response.status === 200 || response.status === 302) {
+        if (!popup) {
+          alert("팝업 차단을 해제해주세요")
+        } else {
+          popup.location.href = response.request?.responseURL;
+        }
+      }
+       else {
+        const data = response.json();
+        data.then((res => {alert(res.errorName)}));
+      }
+    })
+    .catch(error => {
+      console.error(error);
+    });
+  }
+
+  const accessToken = getAccessToken()
+  if (accessToken) {
+    fetchUserDeliveryInfo(accessToken, orderId);
+  } else {
+    if (phoneNumber) {
+      fetchGuestDeliveryInfo(phoneNumber, orderDetails.orderNum);
+    } else {
+      const enteredPhoneNumber = prompt("휴대폰 번호를 입력해주세요.");
+      fetchGuestDeliveryInfo(orderId, enteredPhoneNumber);
+    }
+  }
+}
